Introduce the new KeyManagment and PreKeys implementation
The PreKey class now deals with key creation and verification only, does not store the
key data, etc. The KeyManagement class handles the key data in the database and also
calls the server API implementation to update the key data on the key server.

The KeyManagement caller can now use different implementations of the server APIs.